refactor: extract compaction and lock helpers out of conversation.ts
